# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Return API
  version: 1.1.0
  description: |
    **Return** — API de gestion de prêts d'objets personnels et d'argent.

    Cette API permet de :
    - Gérer des prêts d'objets (photo, retour physique) et d'argent (montant, remboursement)
    - Suivre le cycle de vie complet d'un prêt (création, confirmation, rappels, retour)
    - Planifier et envoyer des rappels automatiques diplomatiques (J-3, J, J+7, J+14, J+21)
    - Consulter l'historique et les statistiques de prêts

    **Architecture** : Monolithe Modulaire (NestJS 11.x + PostgreSQL 17+)
    **Authentification** : JWT (Bearer Token) avec révocation immédiate via Redis
    **Standards** : RFC 7807 pour toutes les erreurs, i18n FR/EN
    **Validation** : class-validator + class-transformer

  contact:
    name: Return Support
    email: support@return.app
    url: https://return.app/support
  license:
    name: Proprietary
    url: https://return.app/license
servers:
  - url: https://api.return.app/v1
    description: Production (Fly.io — Europe West)
  - url: http://localhost:3000/v1
    description: Développement local
...
security:
  - BearerAuth: []
paths:
...
  /borrowers:
    get:
      tags:
        - Borrowers
      summary: Lister tous les contacts (emprunteurs)
      description: |
        Récupère la liste des emprunteurs avec leurs statistiques de fiabilité.
      operationId: listBorrowers
      parameters:
        - name: sortBy
          in: query
          description: Champ de tri
          schema:
            type: string
            enum: [firstName, lastName, trustScore, totalLoans]
            default: firstName
          example: trustScore
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: page
          in: query
          description: Numéro de page (commence à 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Nombre de résultats par page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Liste des emprunteurs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Borrower'
                  pagination:
                    $ref: '#/components/schemas/PaginationMetadata'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Borrowers
      summary: Ajouter un nouveau contact
      description: Créer un emprunteur sans prêt associé (carnet d'adresses).
      operationId: createBorrower
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBorrowerDto'
      responses:
        '201':
          description: Contact créé avec succès
          headers:
            Location:
              schema:
                type: string
                format: uri
                example: https://api.return.app/v1/borrowers/5d6e7f8a-1b2c-4d3e-a5f6-7a8b9c0d1e2f
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Borrower'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Contact déjà existant avec cet email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://api.return.app/errors/borrower-already-exists
                title: Borrower Already Exists
                status: 409
                detail: A contact with this email already exists in your list.
                instance: /v1/borrowers
                timestamp: 2026-02-08T19:00:00Z
                requestId: req-2f3a4b5c
  /borrowers/{borrowerId}:
    parameters:
      - name: borrowerId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Borrowers
      summary: Récupérer les détails d'un emprunteur
      description: Retourne les informations d'un emprunteur ainsi que ses statistiques de fiabilité.
      operationId: getBorrowerById
      responses:
        '200':
          description: Détails de l'emprunteur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Borrower'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Borrowers
      summary: Mettre à jour un contact
      description: Permet de modifier les informations d'un emprunteur (nom, email, téléphone).
      operationId: updateBorrower
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  minLength: 1
                  maxLength: 50
                lastName:
                  type: string
                  minLength: 1
                  maxLength: 50
                email:
                  type: string
                  format: email
                phoneNumber:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Contact mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Borrower'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Contact déjà existant avec cet email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://api.return.app/errors/borrower-already-exists
                title: Borrower Already Exists
                status: 409
                detail: A contact with this email already exists in your list.
                instance: /v1/borrowers/5d6e7f8a-1b2c-4d3e-a5f6-7a8b9c0d1e2f
                timestamp: 2026-02-08T19:00:00Z
                requestId: req-3f4a5b6c

    delete:
      tags:
        - Borrowers
      summary: Supprimer un contact
      description: |
        Supprime un emprunteur s'il n'a aucun prêt actif.
        L'historique des prêts passés est préservé.
      operationId: deleteBorrower
      responses:
        '204':
          description: Contact supprimé avec succès
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: L'emprunteur a des prêts actifs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /borrowers/{borrowerId}/statistics:
    parameters:
      - name: borrowerId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Borrowers
      summary: Récupérer les statistiques de fiabilité d'un emprunteur
      description: Retourne les statistiques détaillées de fiabilité d'un emprunteur, incluant le score de confiance et les délais moyens de retour.
      operationId: getBorrowerStatistics
      responses:
        '200':
          description: Statistiques détaillées
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowerStatistics'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT obtenu via `/auth/login` ou `/auth/register`.
        Format : `Authorization: Bearer <access_token>`

        Le token expire après 15 minutes. Utiliser `/auth/refresh` pour le renouveler.
  schemas:
...
    Borrower:
      type: object
      description: |
        Personne physique à qui un objet ou de l'argent est prêté.
        Doit posséder un compte Return (inscription obligatoire).
        Si le compte n'existe pas encore lors de la création du prêt, une invitation
        est envoyée par email. Le prêt reste en PENDING_CONFIRMATION jusqu'à inscription.
      required:
        - id
        - firstName
        - lastName
        - email
      properties:
        id:
          type: string
          format: uuid
          example: 5d6e7f8a-1b2c-4d3e-a5f6-7a8b9c0d1e2f
        firstName:
          type: string
          minLength: 1
          maxLength: 50
          example: Marie
        lastName:
          type: string
          minLength: 1
          maxLength: 50
          example: Dupont
        email:
          type: string
          format: email
          example: marie.dupont@example.com
          description: Email de l'emprunteur (obligatoire, sert d'identifiant pour le lien de compte)
        phoneNumber:
          type: string
          nullable: true
          example: '+33612345678'
          description: Numéro de téléphone (chiffré en base)
        userId:
          type: string
          format: uuid
          nullable: true
          example: 9a8b7c6d-5e4f-4a3b-8c2d-1e0f9a8b7c6d
          description: ID utilisateur Return (null si invitation en attente)
        statistics:
          $ref: '#/components/schemas/BorrowerStatistics'
    BorrowerStatistics:
      type: object
      description: Statistiques de fiabilité d'un emprunteur
      properties:
        totalLoans:
          type: integer
          example: 12
          description: Nombre total de prêts reçus
        returnedOnTime:
          type: integer
          example: 10
          description: Nombre de prêts rendus à temps
        returnedLate:
          type: integer
          example: 1
          description: Nombre de prêts rendus en retard
        notReturned:
          type: integer
          example: 1
          description: Nombre de prêts non rendus (abandonné)
        averageReturnDelay:
          type: integer
          nullable: true
          example: 2
          description: Délai moyen de retour en jours (+ = retard, - = avance, null si aucun prêt rendu)
        trustScore:
          type: number
          format: float
          minimum: 0
          maximum: 100
          example: 87.5
          description: |
            Score de confiance calculé : 
            (returnedOnTime * 100 + returnedLate * 50) / totalLoans
...
    PaginationMetadata:
      type: object
      description: Métadonnées de pagination pour les listes paginées
      required:
        - currentPage
        - totalPages
        - totalItems
        - itemsPerPage
      properties:
        currentPage:
          type: integer
          example: 1
          description: Page actuelle (commence à 1)
        totalPages:
          type: integer
          example: 5
          description: Nombre total de pages
        totalItems:
          type: integer
          example: 87
          description: Nombre total d'éléments
        itemsPerPage:
          type: integer
          example: 20
          description: Nombre d'éléments par page
        hasNextPage:
          type: boolean
          example: true
          description: Y a-t-il une page suivante ?
        hasPreviousPage:
          type: boolean
          example: false
          description: Y a-t-il une page précédente ?
...
    CreateBorrowerDto:
      type: object
      description: |
        Ajouter un emprunteur à son carnet de contacts.
        L'email est obligatoire car il sert d'identifiant pour le lien vers le compte Return.
        Si l'emprunteur n'a pas encore de compte, une invitation lui sera envoyée.
      required:
        - firstName
        - lastName
        - email
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 50
          example: Marie
        lastName:
          type: string
          minLength: 1
          maxLength: 50
          example: Dupont
        email:
          type: string
          format: email
          example: marie.dupont@example.com
          description: Email obligatoire (identifiant de compte Return)
        phoneNumber:
          type: string
          nullable: true
          example: '+33612345678'
...
    ProblemDetails:
      type: object
      description: |
        Format standardisé RFC 7807 pour toutes les erreurs API.
        Documentation : https://tools.ietf.org/html/rfc7807
      required:
        - type
        - title
        - status
        - detail
        - instance
        - timestamp
        - requestId
      properties:
        type:
          type: string
          format: uri
          example: https://api.return.app/errors/loan-not-found
          description: URI identifiant le type d'erreur
        title:
          type: string
          example: Loan Not Found
          description: Titre court lisible par un humain
        status:
          type: integer
          example: 404
          description: Code HTTP de l'erreur
        detail:
          type: string
          example: "The loan with ID 'loan-123' does not exist or you don't have access to it."
          description: Explication détaillée de l'erreur
        instance:
          type: string
          example: /v1/loans/loan-123
          description: URI de la requête ayant causé l'erreur
        timestamp:
          type: string
          format: date-time
          example: '2026-02-08T14:32:00Z'
          description: Date/heure de l'erreur
        requestId:
          type: string
          example: req-7f3c9a2b-4d1e-4a8f-9c7b-1e3f5a6b8c9d
          description: ID unique de la requête (pour traçabilité)
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          description: Liste des erreurs de validation (optionnel)
    ErrorDetail:
      type: object
      description: Détail d'une erreur de validation sur un champ spécifique
      required:
        - field
        - code
        - message
      properties:
        field:
          type: string
          example: returnDate
          description: Nom du champ en erreur
        code:
          type: string
          example: INVALID_DATE
          description: Code d'erreur (utilisable pour i18n)
        message:
          type: string
          example: Return date must be in the future
          description: Message d'erreur lisible

  # ===========================================================================
  # REUSABLE RESPONSES
  # ===========================================================================
  responses:
    Unauthorized:
      description: Non authentifié (token manquant ou invalide)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            missingToken:
              value:
                type: https://api.return.app/errors/unauthorized
                title: Unauthorized
                status: 401
                detail: No authorization token was found in the request headers.
                instance: /v1/loans
                timestamp: 2026-02-08T14:32:00Z
                requestId: req-9b8c7d6e
            invalidToken:
              value:
                type: https://api.return.app/errors/invalid-token
                title: Invalid Token
                status: 401
                detail: The provided token is invalid or has expired.
                instance: /v1/loans
                timestamp: 2026-02-08T14:35:00Z
                requestId: req-1a2b3c4d
    Forbidden:
      description: Non autorisé (permissions insuffisantes)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            notOwner:
              value:
                type: https://api.return.app/errors/forbidden
                title: Forbidden
                status: 403
                detail: You can only access or modify loans you created.
                instance: /v1/loans/loan-123
                timestamp: 2026-02-08T14:40:00Z
                requestId: req-5e6f7a8b
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            loanNotFound:
              value:
                type: https://api.return.app/errors/loan-not-found
                title: Loan Not Found
                status: 404
                detail: "The loan with ID 'loan-999' does not exist."
                instance: /v1/loans/loan-999
                timestamp: 2026-02-08T14:45:00Z
                requestId: req-9c8d7e6f
    ValidationError:
      description: Erreur de validation des données d'entrée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            multipleErrors:
              value:
                type: https://api.return.app/errors/validation-failed
                title: Validation Failed
                status: 400
                detail: The request contains invalid data.
                instance: /v1/auth/register
                timestamp: 2026-02-08T14:50:00Z
                requestId: req-1f2e3d4c
                errors:
                  - field: email
                    code: INVALID_EMAIL
                    message: Email format is invalid
                  - field: password
                    code: TOO_SHORT
                    message: Password must be at least 8 characters
...